Convert to xml format, jspwiki version stopped forrest build
The jspwiki-parser really, really dislikes lines with repeated "--"
instances.

The line that halted the build was:
{{./configure --without-xfst --with-hfst --enable-spellers --disable-
syntax --disable-transcriptors --disable-analysers --disable-generators
--enable-alignment}}

I tested adding one -- combination at a time.
The time for building the page goes up exponentially as more --
instances are added to a line. On my machine, a couple is ok, the third
one increases the build time from a few seconds to sixteen. Adding a
fourth one lasted so long that I stopped forrest.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/techdoc@131198 c7155fb1-f0a7-4240-a2fc-2600b6f42f90

Former-commit-id: 19a52aca6725c6f2507dd79e20faf447334d7740